use CompositItemProcessor insted of own hack
the self implemented solution using a Groovy Closure was buggy. All
processors always processed the first item istead of the return value of
the privious processor. This only seamed to work correctly because each
processor always returned the same object that was handed to it. But
this is not how it is supposed to be.